changed upload progress block types from NSInteger into int64_t to match STHTTPRequest changes

+++ STTwitter/STHTTPRequest+STTwitter.h
@@ -14,8 +14,8 @@
 + (STHTTPRequest *)twitterRequestWithURLString:(NSString *)urlString
                                     HTTPMethod:(NSString *)HTTPMethod
                               timeoutInSeconds:(NSTimeInterval)timeoutInSeconds
-                  stTwitterUploadProgressBlock:(void(^)(NSInteger bytesWritten, NSInteger totalBytesWritten, NSInteger totalBytesExpectedToWrite))uploadProgressBlock
-                stTwitterDownloadProgressBlock:(void(^)(NSData *data, NSUInteger totalBytesReceived, long long totalBytesExpectedToReceive))stTwitterDownloadProgressBlock
+                  stTwitterUploadProgressBlock:(void(^)(int64_t bytesWritten, int64_t totalBytesWritten, int64_t totalBytesExpectedToWrite))uploadProgressBlock
+                stTwitterDownloadProgressBlock:(void(^)(NSData *data, int64_t totalBytesReceived, int64_t totalBytesExpectedToReceive))stTwitterDownloadProgressBlock
                          stTwitterSuccessBlock:(void(^)(NSDictionary *requestHeaders, NSDictionary *responseHeaders, id json))successBlock
                            stTwitterErrorBlock:(void(^)(NSDictionary *requestHeaders, NSDictionary *responseHeaders, NSError *error))errorBlock;
