+++ STTwitter/STTwitterOSRequest.m
@@ -15,7 +15,7 @@
 
 typedef void (^completion_block_t)(NSObject<STTwitterRequestProtocol> *request, NSDictionary *requestHeaders, NSDictionary *responseHeaders, id response);
 typedef void (^error_block_t)(NSObject<STTwitterRequestProtocol> *request, NSDictionary *requestHeaders, NSDictionary *responseHeaders, NSError *error);
-typedef void (^upload_progress_block_t)(NSInteger bytesWritten, NSInteger totalBytesWritten, NSInteger totalBytesExpectedToWrite);
+typedef void (^upload_progress_block_t)(int64_t bytesWritten, int64_t totalBytesWritten, int64_t totalBytesExpectedToWrite);
 typedef void (^stream_block_t)(NSObject<STTwitterRequestProtocol> *request, NSData *data);
 
 @interface STTwitterOSRequest ()
@@ -34,7 +34,6 @@ typedef void (^stream_block_t)(NSObject<STTwitterRequestProtocol> *request, NSDa
 @property (nonatomic) NSTimeInterval timeoutInSeconds;
 @end
 
-
 @implementation STTwitterOSRequest
 
 - (instancetype)initWithAPIResource:(NSString *)resource
@@ -43,7 +42,7 @@ typedef void (^stream_block_t)(NSObject<STTwitterRequestProtocol> *request, NSDa
                          parameters:(NSDictionary *)params
                             account:(ACAccount *)account
                    timeoutInSeconds:(NSTimeInterval)timeoutInSeconds
-                uploadProgressBlock:(void(^)(NSInteger bytesWritten, NSInteger totalBytesWritten, NSInteger totalBytesExpectedToWrite))uploadProgressBlock
+                uploadProgressBlock:(void(^)(int64_t bytesWritten, int64_t totalBytesWritten, int64_t totalBytesExpectedToWrite))uploadProgressBlock
                         streamBlock:(void(^)(NSObject<STTwitterRequestProtocol> *request, NSData *data))streamBlock
                     completionBlock:(void(^)(NSObject<STTwitterRequestProtocol> *request, NSDictionary *requestHeaders, NSDictionary *responseHeaders, id response))completionBlock
                          errorBlock:(void(^)(NSObject<STTwitterRequestProtocol> *request, NSDictionary *requestHeaders, NSDictionary *responseHeaders, NSError *error))errorBlock {
